Split the three custom-op modules into facade + guarded implementation
Completes the import-safety sweep: `python/cudnn` now has no runtime PyTorch
import outside `cudnn/_deps/torch_dep.py`, and the audit passes clean.

- `ops/causal_conv1d.py`, `experimental/ops/sdpa.py` and
  `experimental/ops/moe_grouped_matmul.py` become import-safe facades that
  resolve their public callables on first attribute access via module
  `__getattr__`, preserving the original function objects and therefore their
  signatures.
- The former module bodies move to `_causal_conv1d_torch.py`, `_sdpa_torch.py`
  and `_moe_grouped_matmul_torch.py`, each wrapped in a single top-level
  `if torch_dep.is_available():` block that binds torch through
  `torch_dep.require()`. Because Python caches modules, the operator schemas,
  fake implementations, compiler hooks and autograd formulas inside still
  register exactly once when PyTorch is present, with unchanged operator names.
  These modules keep real (non-postponed) annotations, which
  `torch.library.custom_op` needs for schema inference.
- `ops/__init__.py` and `experimental/ops/__init__.py` resolve names lazily and
  deliberately bind from the implementation modules, not the same-named facade
  submodules: importing e.g. `cudnn.ops.causal_conv1d` sets it as an attribute
  of the parent package and would otherwise shadow the function of the same
  name.

Adds `test/python/tools/smoke_no_torch_import.py`, which blocks torch with a
MetaPathFinder raising ModuleNotFoundError for exactly `torch`/`torch.*`, then
imports every submodule and asserts the representative entry points raise
`TorchNotAvailableError`.

Verified with torch absent and the pybind module stubbed: `import cudnn` does not
import torch, 212 submodules import cleanly, none hard-require torch, and
`cudnn.Graph`, the causal-conv1d names and both experimental ops all raise
`TorchNotAvailableError` with the canonical message.
